+++ src/pages/home/Home.tsx
@@ -1,4 +1,4 @@
-import { Cards } from '../../components/cardsItems/Cards';
+import { Cards } from '../../components/cards/Cards';
 import { Search } from '../../components/search/Search';
 import './home.css';
